switched from float to double

+++ colour.h
@@ -1,14 +1,14 @@
 #ifndef COLOUR_H
 #define COLOUR_H
 
-#include "float3.h"
+#include "double3.h"
 
-using colour = float3;
+using colour = double3;
 
 void write_colour(std::ostream& out, const colour& pixel_colour) {
-    float r = pixel_colour.x();
-    float g = pixel_colour.y();
-    float b = pixel_colour.z();
+    double r = pixel_colour.x();
+    double g = pixel_colour.y();
+    double b = pixel_colour.z();
 
     // translate [0,1] to byte range [0,255]
     int rbyte = int(255.999f * r);
